Contract_Analysis/requirements.txt
# Core UI
streamlit>=1.32.0

# NLP & Language Detection
spacy>=3.7.2
langdetect>=1.0.9
nltk>=3.8.1

# Hindi → English Offline Translation
argostranslate>=1.9.1

# Document Reading
pdfplumber>=0.10.3
python-docx>=1.1.0

# PDF Report Generation
reportlab>=4.0.9

# Utilities
numpy>=1.26.0
scikit-learn>=1.4.0

# spaCy English Model (install separately)
# python -m spacy download en_core_web_sm
